neovim: minor improvements and plugin updates

+++ xdg_config/nvim/plugin/commands.lua
@@ -1,7 +1,7 @@
 vim.api.nvim_create_user_command("Del", function(args)
   local bufnr = vim.api.nvim_get_current_buf()
   local fname = vim.api.nvim_buf_get_name(bufnr)
-  if vim.bo[bufnr].buftype == "" then
+  if vim.bo[bufnr].buftype == "" and fname ~= "" then
     local ok, err = os.remove(fname)
     assert(args.bang or ok, err)
   end
